asterra/geojson_loader/requirements.txt
flask==3.1.1
boto3==1.40.0
psycopg2-binary==2.9.10
geojson==3.2.0
geopandas==0.14.3
watchtower==3.0.1
pytest==8.2.2
